requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
django-braces==1.15.0
django-ckeditor==6.5.1
django-comments-dab==2.8.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-js-asset==2.0.0
django-model-utils==4.3.1
django-social-share==2.3.0
django-star-ratings==0.9.2
django-widget-tweaks==1.4.12
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
swapper==1.3.0
tzdata==2022.7
urllib3==1.26.13
